spark/requirements.txt
py4j==0.10.9.5
cassandra-driver
kafka-python
minio
pyspark
python-dateutil
datetime
